requirements.txt
streamlit>=1.36
numpy>=1.26
sentence-transformers>=3.0
faiss-cpu>=1.8

